pilotd/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pilotd"
version = "0.0.1"
description = "Pilot daemon — observe/think/act agent and socket server"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "apscheduler>=3.10.4",
    "sqlalchemy>=2.0.0",
    "sqlite-vec>=0.1.6",
    "onnxruntime>=1.20.0",
    "pyobjc-framework-Quartz>=10.3",
    "pyobjc-framework-Cocoa>=10.3",
    "pyobjc-framework-ApplicationServices>=10.3",
    "Pillow>=10.4.0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pilot*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
